requirements.txt
flask==3.0.0
schedule==1.2.1
line-bot-sdk==3.*
python-dotenv==1.0.1
